requirements.txt
pyrogram
tgcrypto
google-generativeai
python-dotenv
